fix(enforcers): verbose option validation
Fix the verbose validator so it doesn't throw when no value is provided. The `osom` library runs the validate function regardless of whether the prop was even provided, so that validator has been removed. There's already a check in the query runner anyway as to whether `instance.verbose` is a function or not.