refactor: revise some error messages, add note to `create()`
Make a note in `create()` about the seemingly redundant existence check on the table. It prevents an error from being thrown when inserting a table that already exists due to a unique index already being declared. Using `createTableIfNotExists` does not prevent this error.

Modify a few error messages for style and clarity.

Rearrange if order in `isValidWhere` for potential perf differences.