feat(create): return the created object
Change the return value of `create()` to be the newly created object, or the existing one if a unique violation was encountered.

This makes more sense given that in general the number modifed would always have been 0 or 1 so it might as well give you the object if 1, or nothing if 0.

BREAKING CHANGE: The return value of `create()` is no longer the number of created objects. Instead, `create()` returns the created object.